Update youtube script for API v3
Starting April 20th, Google are shutting down the old v2 API.
As a result, hubot sometimes returns a link to a "device support" video instead of a real result:
https://www.youtube.com/watch?v=UKY3scPIMd8

This upgrades to v3 and uses the new search endpoint in place of the v2 videos one which no longer takes a keyword filter.
The data returned by the API also no longer includes the video link, so we form it ourselves instead.

Also adds some tweaks made to Liobot a while ago that add "norandom" support. By default it'll return a random video unless the query is prefixed with a dash.